新增 新增客戶console

+++ app/Services/RankService.php
@@ -8,6 +8,7 @@ use App\Repositories\RankRepository;
 use App\Helpers\CustomerHelper;
 use App\Exceptions\ResourceNotFoundException;
 use Cache;
+use Symfony\Component\HttpKernel\Exception\ConflictHttpException;
 
 class RankService
 {
@@ -50,6 +51,9 @@ class RankService
     public function deleteRank($id)
     {
         $rank = $this->findRank($id);
+        if ($rank->members->count() > 0) {
+            throw new ConflictHttpException('The rank has members. Can not be deleted.');
+        }
         $rank->delete();
         return $rank;
     }
